Move ParallelControl to digdag-plugin-utils project
